Build mantohtml with the build architecture compiler
mantohtml is run during build. Thus it needs to be built with the build
architecture compiler (or execution fails). The obvious part is switching to
CC_FOR_BUILD. That also depends on it not requiring any other cups components.
In particular, removing uses of strlcpy and replacing host architecture-
specific includes is thus needed.

Bug-Debian: https://bugs.debian.org/837936

Gbp-Pq: Name 0037-Build-mantohtml-with-the-build-architecture-compiler.patch