PPD Generator for driverless printing setup: Workaround for wrong resolution.
There are some PWG Raster printers with a firmware bug which report a
wrong resolution as pwg-raster-document-resolution-supported IPP
attribute. If this happens (horizontal and/or vertical resolution is <
75 dpi) we try the printer-resolution-supported attribute and if this
is also wrong fall back to 300 dpi.

Passed over from cups-filters, which contains a copy of CUPS' PPD
generator due to the fact that CUPS did not put it into its public
API.

Bug-Debian: https://bugs.debian.org/868360
Patch-Name: pwg-raster-resolution-fallback.patch