Fix potential unaligned accesses in the string pool (Issue #5474)
This set of changes makes the PPD functions use strdup and free - they were
modifying the contents of the string in places and doing other things that
were not safe for (immutable) strings in the pool.

Clean out some more _cupsStr cruft that might potentially cause an unaligned memory access (Issue #5474)

Don't directly use the string pool in the CGI programs or scheduler.